requirements.txt
django-widgets-improved==1.5.0